+++ docs/source/whats_new.rst
@@ -20,7 +20,7 @@ Develop branch
 Enhancements
 ~~~~~~~~~~~~
 
-
+- Introduce decoupled scaling factors for GrowingModule behavior, separating scaling for the optimal-delta update, the input extension, and the output extension (:gh:`245` by `Théo Rudkiewicz`_).
 - Introduce VGG-like growing architecture support with the `VGG` class (:gh:`236` by `Hugo Mousset`_)
 - Adds support for selecting GroupNorm as the normalization layer in the ResNet container, alongside the existing BatchNorm option (:gh:`233` by `Théo Rudkiewicz`_).
 - Adds configurability to the ResNet container to allow using BatchNorm2d or disabling normalization entirely (:gh:`228` by `Théo Rudkiewicz`_)